+++ sdk/core/azure-core-experimental/src/main/java/com/azure/core/experimental/serializer/JsonSerializer.java
@@ -3,6 +3,7 @@
 
 package com.azure.core.experimental.serializer;
 
+import com.azure.core.util.serializer.TypeReference;
 import reactor.core.publisher.Mono;
 
 import java.io.InputStream;
@@ -11,29 +12,7 @@ import java.io.OutputStream;
 /**
  * Generic interface covering basic JSON serialization and deserialization methods.
  */
-public interface JsonSerializer extends ObjectSerializer {
-    /**
-     * Reads a JSON stream into its object representation.
-     *
-     * @param stream JSON stream.
-     * @param typeReference {@link TypeReference} representing the object.
-     * @param <T> Type of the object.
-     * @return The object represented by the deserialized JSON stream.
-     */
-    @Override
-    <T> T deserialize(InputStream stream, TypeReference<T> typeReference);
-
-    /**
-     * Reads a JSON stream into its object representation.
-     *
-     * @param stream JSON stream.
-     * @param typeReference {@link TypeReference} representing the object.
-     * @param <T> Type of the object.
-     * @return The object represented by the deserialized JSON stream.
-     */
-    @Override
-    <T> Mono<T> deserializeAsync(InputStream stream, TypeReference<T> typeReference);
-
+public interface JsonSerializer extends com.azure.core.util.serializer.JsonSerializer {
     /**
      * Reads a JSON tree into its object representation.
      *
@@ -53,28 +32,6 @@ public interface JsonSerializer extends ObjectSerializer {
      * @return The object represented by the deserialized JSON tree.
      */
     <T> Mono<T> deserializeTreeAsync(JsonNode jsonNode, TypeReference<T> typeReference);
-
-    /**
-     * Writes an object's JSON into a stream..
-     *
-     * @param stream {@link OutputStream} where the object's JSON will be written.
-     * @param value The object.
-     * @param <S> Type of the output stream.
-     * @return The stream where the object's JSON was written.
-     */
-    @Override
-    <S extends OutputStream> S serialize(S stream, Object value);
-
-    /**
-     * Writes an object's JSON into a stream..
-     *
-     * @param stream {@link OutputStream} where the object's JSON will be written.
-     * @param value The object.
-     * @param <S> Type of the output stream.
-     * @return The stream where the object's JSON was written.
-     */
-    @Override
-    <S extends OutputStream> Mono<S> serializeAsync(S stream, Object value);
 
     /**
      * Writes a JSON tree into a stream.
