+++ sdk/core/azure-core-experimental/src/main/java/module-info.java
@@ -1,12 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+import com.azure.core.experimental.serializer.AvroSerializerProvider;
+
 module com.azure.core.experimental {
     requires transitive com.azure.core;
 
     exports com.azure.core.experimental.serializer;
     exports com.azure.core.experimental.spatial;
 
-    uses com.azure.core.experimental.serializer.AvroSerializerProvider;
-    uses com.azure.core.experimental.serializer.JsonSerializerProvider;
+    uses AvroSerializerProvider;
 }
