+++ sdk/core/azure-core-serializer-avro-apache/src/main/resources/META-INF/services/com.azure.core.experimental.serializer.AvroSerializerProvider
