+++ sdk/core/azure-core-serializer-avro-apache/src/test/java/com/azure/core/serializer/avro/apache/ApacheAvroSerializerTests.java
@@ -3,11 +3,11 @@
 
 package com.azure.core.serializer.avro.apache;
 
-import com.azure.core.experimental.serializer.TypeReference;
 import com.azure.core.serializer.avro.apache.generatedtestsources.HandOfCards;
 import com.azure.core.serializer.avro.apache.generatedtestsources.LongLinkedList;
 import com.azure.core.serializer.avro.apache.generatedtestsources.PlayingCard;
 import com.azure.core.serializer.avro.apache.generatedtestsources.PlayingCardSuit;
+import com.azure.core.util.serializer.TypeReference;
 import org.apache.avro.util.Utf8;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
