+++ sdk/core/azure-core-serializer-avro-jackson/src/main/resources/META-INF/services/com.azure.core.experimental.serializer.AvroSerializerProvider
