+++ sdk/core/azure-core-serializer-json-gson/pom.xml
@@ -60,11 +60,6 @@
       <artifactId>azure-core</artifactId>
       <version>1.7.0-beta.3</version> <!-- {x-version-update;com.azure:azure-core;current} -->
     </dependency>
-    <dependency>
-      <groupId>com.azure</groupId>
-      <artifactId>azure-core-experimental</artifactId>
-      <version>1.0.0-beta.2</version> <!-- {x-version-update;com.azure:azure-core-experimental;current} -->
-    </dependency>
 
     <dependency>
       <groupId>com.google.code.gson</groupId>
