+++ sdk/core/azure-core-serializer-json-gson/src/main/java/module-info.java
@@ -3,11 +3,10 @@
 
 module com.azure.core.serializer.json.gson {
     requires transitive com.azure.core;
-    requires transitive com.azure.core.experimental;
     requires transitive gson;
 
     exports com.azure.core.serializer.json.gson;
 
-    provides com.azure.core.experimental.serializer.JsonSerializerProvider
+    provides com.azure.core.util.serializer.JsonSerializerProvider
         with com.azure.core.serializer.json.gson.GsonJsonSerializerProvider;
 }
