+++ sdk/core/azure-core-serializer-json-gson/src/main/resources/META-INF/services/com.azure.core.util.serializer.JsonSerializerProvider
