+++ sdk/core/azure-core-serializer-json-jackson/src/main/java/module-info.java
@@ -3,10 +3,9 @@
 
 module com.azure.core.serializer.json.jackson {
     requires transitive com.azure.core;
-    requires transitive com.azure.core.experimental;
 
     exports com.azure.core.serializer.json.jackson;
 
-    provides com.azure.core.experimental.serializer.JsonSerializerProvider
+    provides com.azure.core.util.serializer.JsonSerializerProvider
         with com.azure.core.serializer.json.jackson.JacksonJsonSerializerProvider;
 }
