+++ sdk/core/azure-core-serializer-json-jackson/src/main/resources/META-INF/services/com.azure.core.util.serializer.JsonSerializerProvider
