+++ sdk/core/azure-core/src/main/java/module-info.java
@@ -41,8 +41,9 @@ module com.azure.core {
     exports com.azure.core.implementation.http to com.azure.core.test;
 
     // Service Provider Interfaces
-    uses com.azure.core.util.tracing.Tracer;
     uses com.azure.core.http.HttpClientProvider;
     uses com.azure.core.http.policy.BeforeRetryPolicyProvider;
     uses com.azure.core.http.policy.AfterRetryPolicyProvider;
+    uses com.azure.core.util.serializer.JsonSerializerProvider;
+    uses com.azure.core.util.tracing.Tracer;
 }
